Exclude health from logs in orchestrator + refactor middlewares
