+++ packages/api/internal/auth/constants.go
@@ -1,4 +1,4 @@
-package constants
+package auth
 
 const TeamContextKey string = "team"
 const UserIDContextKey string = "user_id"
