+++ packages/api/internal/handlers/sandbox_kill.go
@@ -2,9 +2,9 @@ package handlers
 
 import (
 	"fmt"
+	"github.com/e2b-dev/infra/packages/api/internal/auth"
 	"net/http"
 
-	"github.com/e2b-dev/infra/packages/api/internal/constants"
 	"github.com/e2b-dev/infra/packages/api/internal/utils"
 	"github.com/e2b-dev/infra/packages/shared/pkg/models"
 	"github.com/e2b-dev/infra/packages/shared/pkg/telemetry"
@@ -20,7 +20,7 @@ func (a *APIStore) DeleteSandboxesSandboxID(
 	ctx := c.Request.Context()
 	sandboxID = utils.ShortID(sandboxID)
 
-	teamID := c.Value(constants.TeamContextKey).(models.Team).ID
+	teamID := c.Value(auth.TeamContextKey).(models.Team).ID
 
 	telemetry.SetAttributes(ctx,
 		attribute.String("instance.id", sandboxID),
