+++ packages/shared/pkg/gin_utils/middleware/otel/metrics/option.go
